requirements.txt
numpy
pandas
matplotlib
pyyaml
scipy
pytest
pygad